requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.1.0
pytest==8.3.5
email-validator==2.2.0


